+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.13.2 (unreleased)
 
 - Optimize index size by removing ranges that have codepoints embedded
+- Optimize index size by substituting common words
 - Fix missing Tangut ideographs
 
 ### 1.13.1